move setupPackageManager to a new file
